Add support for failing a request with http status codes instead of just 400 bad request (useful for OnAuthorizeAsync)

+++ Source/tusdotnet/ExternalMiddleware/Core/TusCoreMiddleware.cs
@@ -58,7 +58,7 @@ namespace tusdotnet
             {
                 Body = context.Response.Body,
                 SetHeader = (key, value) => context.Response.Headers[key] = value,
-                SetStatus = status => context.Response.StatusCode = status
+                SetStatus = status => context.Response.StatusCode = (int)status
             };
 
             var handled = await TusProtocolHandlerIntentBased.Invoke(new ContextAdapter
